commit 19

+++ Hangman/Frames/keyboard.py
@@ -1,6 +1,8 @@
-import json, os.path
+import json
 import customtkinter as ctk
 from CTkMessagebox import CTkMessagebox
+
+from Utility.static_functions import create_file
 
 
 class KeyboardFrame(ctk.CTkFrame):
@@ -25,8 +27,6 @@ class KeyboardFrame(ctk.CTkFrame):
 
                 self.btn.grid(row=1, column=i - 13)
 
-        # self.status()
-
     def __on_click(self, event):
         char = event.widget.master.cget('text')
 
@@ -45,6 +45,7 @@ class KeyboardFrame(ctk.CTkFrame):
             self.ms.score_frame.set_attempt()
             self.ms.score_frame.set_remaining_tries()
             self.ms.status_frame.update_image()
+
         self.status()
 
     def status(self):
@@ -73,40 +74,32 @@ class KeyboardFrame(ctk.CTkFrame):
 
         elif self.ms.score_frame.get_remaining_tries() == 0:
 
-            message = CTkMessagebox(master=self.ms, title='Exit or Continue', icon='question',
-                                    message='You have no remaining tries!\nStart a new game?',
-                                    width=430, option_1='Yes', option_2='No', cancel_button='No')
+            message = CTkMessagebox(master=self.ms, title='Game Lost', icon='info',
+                                    message='You have no remaining tries!', height=100,
+                                    width=200, option_1='Ok', cancel_button='Ok')
             response = message.get()
 
-            self.add_player_score_to_file()
+            if response == 'Ok':
 
-            if response == 'No':
-
-                self.ms.update()
-
-            else:
-
+                self.add_player_score_to_file()
                 self.ms.score_frame.reset_score()
                 self.ms.category_frame.ls = []
                 self.status()
-                return self.ms.category_frame.refresh_sc_wd_frame()
-
-        # self.after(1000, lambda: self.status())
+                self.ms.category_frame.refresh_sc_wd_frame()
 
     def add_player_score_to_file(self):
         json_file = 'high_scorers.json'
         name = self.ms.top_score_frame.player_label.cget('text')
         score = self.ms.score_frame.get_score()
-        self.__create_file(json_file)
-        with open(json_file, 'r+') as file:
-            score_dict = json.load(file)
-            score_dict[len(score_dict)] = {name: score}
-            file.seek(0)
-            file.truncate()
-            json.dump(score_dict, file, indent=4)
+        create_file(json_file)
 
-    @staticmethod
-    def __create_file(file_name):
-        if not os.path.exists(file_name):
-            with open(file_name, 'x') as create_file:
-                json.dump({}, create_file)
+        if name != 'Guest':
+
+            with open(json_file, 'r+') as file:
+                score_dict = json.load(file)
+                score_dict[len(score_dict)] = {name: score}
+                file.seek(0)
+                file.truncate()
+                json.dump(score_dict, file, indent=4)
+
+
